pyproject.toml
[tool.poetry]
name = "mydeardiary"
version = "0.1.0"
description = ""
authors = ["GodVov4 <volume245@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
django = "4.2.7"
psycopg2-binary = "2.9.9"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
